+++ packages/forklift-console-plugin/src/modules/Plans/views/details/components/UpdateMappings/index.ts
@@ -1,0 +1,6 @@
+// @index(['./*', /style/g], f => `export * from '${f.path}';`)
+export * from './initialPlanMappingsState';
+export * from './PlanMappingsInitSection';
+export * from './PlanMappingsSection';
+export * from './reducer';
+// @endindex
