+++ packages/forklift-console-plugin/src/modules/Plans/views/details/components/UpdateMappings/initialPlanMappingsState.ts
@@ -1,11 +1,13 @@
+import { PlanEditAction } from 'src/modules/Plans/utils/types/PlanEditAction';
+
 import { V1beta1NetworkMap, V1beta1StorageMap } from '@kubev2v/types';
 
-import { PlanEditAction } from '../../utils/types/PlanEditAction';
-import { PlanMappingsSectionState } from '../details';
+import { PlanMappingsSectionState } from './PlanMappingsSection';
 
 export type InitialPlanMappingsStateProps = {
   planNetworkMaps: V1beta1NetworkMap;
   planStorageMaps: V1beta1StorageMap;
+  edit: boolean;
   editAction?: PlanEditAction;
 };
 
@@ -13,8 +15,9 @@ export const initialPlanMappingsState = ({
   planNetworkMaps,
   planStorageMaps,
   editAction,
+  edit,
 }: InitialPlanMappingsStateProps): PlanMappingsSectionState => ({
-  edit: true,
+  edit,
   dataChanged: false,
   alertMessage: null,
   updatedNetwork: planNetworkMaps?.spec?.map || [],
