+++ packages/forklift-console-plugin/src/modules/Plans/views/details/tabs/Mappings/index.ts
@@ -1,4 +1,3 @@
 // @index(['./*', /style/g], f => `export * from '${f.path}';`)
 export * from './PlanMappings';
-export * from './PlanMappingsSection';
 // @endindex
